# src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Genre": {
				"enum": [
					"Fiction",
					"Non-Fiction",
					"Mystery",
					"Sci-Fi",
					"Fantasy",
					"Biography",
					"History",
					"Romance",
					"Thriller",
					"Horror",
					"Poetry",
					"Drama",
					"Comics",
					"Other"
				],
				"type": "string"
			},
			"BookStatus": {
				"enum": [
					"available",
					"borrowed"
				],
				"type": "string"
			},
			"Book": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"author": {
						"type": "string"
					},
					"pages": {
						"type": "number",
						"format": "double"
					},
					"year": {
						"type": "number",
						"format": "double"
					},
					"genre": {
						"$ref": "#/components/schemas/Genre"
					},
					"status": {
						"$ref": "#/components/schemas/BookStatus"
					}
				},
				"required": [
					"id",
					"title",
					"author",
					"pages",
					"year",
					"genre",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"VisitorWithBooks": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"surname": {
						"type": "string"
					},
					"registrationDate": {
						"type": "string"
					},
					"currentBooks": {
						"items": {
							"$ref": "#/components/schemas/Book"
						},
						"type": "array"
					},
					"history": {
						"items": {
							"$ref": "#/components/schemas/Book"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"surname",
					"registrationDate",
					"currentBooks",
					"history"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ErrorResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_VisitorWithBooks-Array_": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/VisitorWithBooks"
						},
						"type": "array"
					},
					"successful": {
						"type": "boolean"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorResponse"
					}
				},
				"required": [
					"successful"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_VisitorWithBooks_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/VisitorWithBooks"
					},
					"successful": {
						"type": "boolean"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorResponse"
					}
				},
				"required": [
					"successful"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Link": {
				"properties": {
					"table": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"table",
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Visitor": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"surname": {
						"type": "string"
					},
					"registrationDate": {
						"type": "string"
					},
					"currentBooks": {
						"items": {
							"$ref": "#/components/schemas/Link"
						},
						"type": "array"
					},
					"history": {
						"items": {
							"$ref": "#/components/schemas/Link"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"surname",
					"registrationDate",
					"currentBooks",
					"history"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_Visitor_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/Visitor"
					},
					"successful": {
						"type": "boolean"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorResponse"
					}
				},
				"required": [
					"successful"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateVisitorRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"surname": {
						"type": "string"
					},
					"registrationDate": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"surname",
					"registrationDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_CreateVisitorRequest_": {
				"properties": {
					"name": {
						"type": "string"
					},
					"surname": {
						"type": "string"
					},
					"registrationDate": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"ApiResponse_void_": {
				"properties": {
					"data": {},
					"successful": {
						"type": "boolean"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorResponse"
					}
				},
				"required": [
					"successful"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "lab4",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/visitors": {
			"get": {
				"operationId": "GetVisitors",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_VisitorWithBooks-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Visitors"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "AddVisitor",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Visitor_"
								}
							}
						}
					}
				},
				"tags": [
					"Visitors"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateVisitorRequest"
							}
						}
					}
				}
			}
		},
		"/visitors/{id}": {
			"get": {
				"operationId": "GetVisitor",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_VisitorWithBooks_"
								}
							}
						}
					}
				},
				"tags": [
					"Visitors"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateVisitor",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_VisitorWithBooks_"
								}
							}
						}
					}
				},
				"tags": [
					"Visitors"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_CreateVisitorRequest_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteVisitor",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_void_"
								}
							}
						}
					}
				},
				"tags": [
					"Visitors"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/visitors/{id}/download": {
			"get": {
				"operationId": "DownloadVisitor",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Visitors"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}